# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kudesnik. Windows \u0026 Doors | REST-API scheme",
        "title": "Kudesnik API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Получить список товаров, есть фильтры",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить список товаров",
                "parameters": [
                    {
                        "type": "string",
                        "description": "код категории товар (windows, entrance-doors, interior-doors, balconies, others)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "лимит возвращаемых категорий",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение возвращаемых категорий",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "сортировка по возрастанию или убыванию (asc, desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBaseDetailed"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить несколько товаров по списку их ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Удалить товары",
                "parameters": [
                    {
                        "description": "DeleteProducts тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_products_transport_http.ProductIdDTORequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/products/{category_code}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать товар (требуется указать конкретную категорию)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Создать товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "код категории товар которой создается (windows, entrance-doors, interior-doors, balconies, others)",
                        "name": "category_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CreateProduct тело запроса, если category_code=others",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный товар",
                        "schema": {
                            "$ref": "#/definitions/internal_features_products_transport_http.CreateProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{category_code}/{id}": {
            "get": {
                "description": "Получить товар (требуется указать конкретную категорию)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить товар",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "код категории товар которой создается (windows, entrance-doors, interior-doors, balconies, others)",
                        "name": "category_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный товар",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductDetailed"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить товар (требуется указать конкретную категорию)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Обновить товар",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "код категории товар которой создается (windows, entrance-doors, interior-doors, balconies, others)",
                        "name": "category_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchProduct тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBasePatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный товар",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBase"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить товар по ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBase": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "producer_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBaseDetailed": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "installation_price": {
                    "type": "number"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "producer_company_name": {
                    "type": "string"
                },
                "producer_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductBasePatch": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "producer_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "thumbnail_url": {
                    "type": "string"
                }
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductDetailed": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/github_com_nikitavaulin_kudesnik_internal_core_domain.ProductDetails"
                },
                "product": {}
            }
        },
        "github_com_nikitavaulin_kudesnik_internal_core_domain.ProductDetails": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "installation_price": {
                    "type": "number"
                },
                "producer_company_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_nikitavaulin_kudesnik_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_products_transport_http.CreateProductResponse": {
            "type": "object",
            "properties": {
                "category_code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "producer_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "internal_features_products_transport_http.ProductIdDTORequest": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Some queries require authorization: \"Bearer \u003cjwt_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}